fix Carrierwave::Storage::Raca::File#read
* this method is (udfortunately unspec'd) because mocking a downloaded
  from raca is a bit tricky
* I've manually confirmed this fix works though

+++ lib/carrierwave/storage/raca.rb
@@ -56,12 +56,12 @@ module CarrierWave
         end
 
         def read
-          tempfile = Tempfile.new
+          tempfile = Tempfile.new("carrierwave-raca")
           tempfile.close
           container.download(path, tempfile.path)
-          File.read(tempfile.path)
+          ::File.read(tempfile.path)
         ensure
-          tempfile.unlink
+          tempfile.unlink if tempfile
         end
 
         def size
